+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * `step_tokenize()` got an `engine` argument to specify packages other then tokenizers to tokenize.
 * `spacyr` have been added as an engine to `step_tokenize()`.
+* `step_lemma()` has been added to extract lemma attribute from tokenlists.
 
 # textrecipes 0.1.0